requirements.txt
chromadb==0.3.20
GitPython==3.1.31
langchain==0.0.131
openai==0.27.3
unstructured==0.5.9
gunicorn==20.1.0
flask==2.2.3
pypdf==3.7.0
python-dotenv~=1.0.0
requests~=2.28.2
colorama~=0.4.6